Share report helpers
