+++ packages/backend/src/services/reports.ts
@@ -1,31 +1,6 @@
 import { prisma } from './db.js';
 import { calcTimeAmount, resolveRateCard } from './rateCard.js';
-
-function toNumber(value: unknown): number {
-  if (value == null) return 0;
-  if (typeof value === 'number') return value;
-  if (typeof value === 'string') {
-    const parsed = Number(value);
-    return Number.isFinite(parsed) ? parsed : 0;
-  }
-  if (value && typeof value === 'object') {
-    const maybeDecimal = value as {
-      toNumber?: () => number;
-      toString?: () => string;
-    };
-    if (typeof maybeDecimal.toNumber === 'function')
-      return maybeDecimal.toNumber();
-    if (typeof maybeDecimal.toString === 'function') {
-      const parsed = Number(maybeDecimal.toString());
-      return Number.isFinite(parsed) ? parsed : 0;
-    }
-  }
-  return 0;
-}
-
-function dateKey(date: Date) {
-  return date.toISOString().slice(0, 10);
-}
+import { dateKey, toNumber } from './utils.js';
 
 async function sumTimeCost(
   projectId: string,
@@ -87,12 +62,18 @@ async function resolveRevenueBudget(
   from?: Date,
   to?: Date,
 ): Promise<number> {
+  const estimateWhere: any = {
+    projectId,
+    deletedAt: null,
+    status: { notIn: ['cancelled', 'rejected'] },
+  };
+  if (from || to) {
+    estimateWhere.createdAt = {};
+    if (from) estimateWhere.createdAt.gte = from;
+    if (to) estimateWhere.createdAt.lte = to;
+  }
   const estimate = await prisma.estimate.findFirst({
-    where: {
-      projectId,
-      deletedAt: null,
-      status: { notIn: ['cancelled', 'rejected'] },
-    },
+    where: estimateWhere,
     orderBy: { createdAt: 'desc' },
     select: { totalAmount: true },
   });
